Add module cache that regroups CacheState, VectorCacheState and CacheImpl
